+++ packages/index-page/src/components/common-utils/bottom-link.js
@@ -4,11 +4,16 @@ import LinkIcon from '../../static/link-arrow.svg'
 import PropTypes from 'prop-types'
 import React from 'react'
 import styled from 'styled-components'
-import color from '../../constants/color'
+import { colorSupportive } from '@twreporter/core/lib/constants/color'
 
+const TextSpan = styled.span``
+const StyledA = styled.a`
+  text-decoration: none;
+`
 const Wrapper = styled.div`
   font-size: 14px;
-  color: ${props => (props.isDarkBg ? color.lightBlue : color.blue)};
+  color: ${props =>
+    props.isDarkBg ? colorSupportive.main : colorSupportive.heavy};
   cursor: pointer;
   display: inline-block;
   > span {
@@ -25,20 +30,15 @@ const Wrapper = styled.div`
       transform: translateX(5px);
     }
   }
-`
-
-const TextSpan = styled.span`
-  color: ${props => {
-    return props.isDarkBg ? color.lightBlue : color.blue
-  }};
-`
-
-const A = styled.a`
-  text-decoration: none;
-  &:visited {
-    color: ${props => {
-      return props.isDarkBg ? color.lightBlue : color.blue
-    }};
+  ${TextSpan} {
+    color: ${props =>
+      props.isDarkBg ? colorSupportive.main : colorSupportive.heavy};
+  }
+  ${StyledA} {
+    &:visited {
+      color: ${props =>
+        props.isDarkBg ? colorSupportive.main : colorSupportive.heavy};
+    }
   }
 `
 
@@ -47,20 +47,15 @@ const BottomLink = props => {
   return (
     <Wrapper isDarkBg={isDarkBg}>
       {redirect ? (
-        <A
-          href={path}
-          target={target}
-          rel="noreferrer noopener"
-          isDarkBg={isDarkBg}
-        >
+        <StyledA href={path} target={target} rel="noreferrer noopener">
           <TextSpan>{text}</TextSpan>
-        </A>
+        </StyledA>
       ) : (
         <Link to={`/${path}`}>
           <TextSpan>{text}</TextSpan>
         </Link>
       )}
-      {isDarkBg ? <LinkIcon /> : <DarkBgIcon />}
+      {isDarkBg ? <DarkBgIcon /> : <LinkIcon />}
     </Wrapper>
   )
 }
@@ -82,4 +77,4 @@ BottomLink.defaultProps = {
 
 export default BottomLink
 
-export { A, TextSpan, Wrapper, LinkIcon, DarkBgIcon }
+export { StyledA, TextSpan, Wrapper, LinkIcon, DarkBgIcon }