+++ packages/react-components/src/confirmation/constants/color.js
@@ -1,6 +1,0 @@
-export default {
-  gray: 'rgba(115, 115, 115, 0.8)',
-  darkGray: '#3e3f3f',
-  lightRed: '#e60013',
-  red: '#c40c23',
-}