+++ packages/react-components/src/error/constants/color.js
@@ -1,4 +1,0 @@
-export default {
-  black: '#000000',
-  white: '#ffffff',
-}