+++ packages/react-components/src/error/message.js
@@ -1,10 +1,9 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import styled from 'styled-components'
-// core
+// @twreporter
 import mq from '@twreporter/core/lib/utils/media-query'
-// constants
-import color from './constants/color'
+import { colorGrayscale } from '@twreporter/core/lib/constants/color'
 // assets
 import Building from './assets/building.svg'
 import Dot from './assets/dot.svg'
@@ -75,14 +74,14 @@ const ChineseText = styled.div`
     font-size: 13px;
     line-height: 1.62;
     text-align: left;
-    color: ${color.black};
+    color: ${colorGrayscale.black};
     font-weight: 900;
     ::after {
       content: "";
       display: block;
       width: 14px;
       height: 1px;
-      background-color: ${color.black};
+      background-color: ${colorGrayscale.black};
       position: relative;
       top: 6px;
       left: 7px;
@@ -95,14 +94,14 @@ const ChineseText = styled.div`
     font-size: 18px;
     line-height: 1.5;
     text-align: left;
-    color: ${color.black};
+    color: ${colorGrayscale.black};
     font-weight: 900;
     ::after {
       content: "";
       display: block;
       width: 20px;
       height: 1px;
-      background-color: ${color.black};
+      background-color: ${colorGrayscale.black};
       position: relative;
       top: 19px;
       left: 11px;
@@ -188,11 +187,11 @@ const BackToHomeBtn = styled.a`
   display: block;
   cursor: pointer;
   text-align: center;
-  background-color: ${color.black};
-  color: ${color.white};
+  background-color: ${colorGrayscale.black};
+  color: ${colorGrayscale.white};
   text-decoration: none;
   &:hover, &:active, &:focus, &:visited {
-    color: ${color.white};
+    color: ${colorGrayscale.white};
     text-decoration: none;
   }
   ${mq.mobileOnly`