+++ packages/redux/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.8.2-rc.2](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/redux@7.8.2-rc.1...@twreporter/redux@7.8.2-rc.2) (2023-12-12)
+
+**Note:** Version bump only for package @twreporter/redux
+
+
+
+
+
+## [7.8.2-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/redux@7.8.2-rc.0...@twreporter/redux@7.8.2-rc.1) (2023-12-05)
+
+**Note:** Version bump only for package @twreporter/redux
+
+
+
+
+
+## [7.8.2-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/redux@7.8.1...@twreporter/redux@7.8.2-rc.0) (2023-12-05)
+
+**Note:** Version bump only for package @twreporter/redux
+
+
+
+
+
 ## [7.8.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/redux@7.8.1-rc.3...@twreporter/redux@7.8.1) (2023-11-20)
 
 **Note:** Version bump only for package @twreporter/redux